backend/requirements.txt
appnope==0.1.0
backcall==0.1.0
boto==2.49.0
boto3==1.9.18
botocore==1.12.18
bz2file==0.98
certifi==2018.8.24
chardet==3.0.4
decorator==4.3.0
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
docutils==0.14
Faker==0.9.1
fast-intensity==0.4
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
jmespath==0.9.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
smart-open==1.7.1
text-unidecode==1.2
tqdm==4.26.0
traitlets==4.3.2
urllib3==1.23
waitress==1.1.0
wcwidth==0.1.7
whitenoise==3.3.1
